frontend/requirements.txt
flet==0.16.0
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.5.3